Run QUnit tests on minified JS file
As Grunt seems to lack the ability to run the same task twice with
different parameters, QUnit now moves to the end of the Grunt build.
Testing the minified JS file provides a way to check that the concat
section of the gruntfile isn't missing any of the .js files.