Rename 'course' to 'class'
This is in preparation for courses being introduced as comprising of
multiple classes, as can be done in the SI format.  The term 'age-class'
is used in various place even in ageless contests.  Some of this is to
avoid variables named 'class', which isn't permitted.  Additionally, the
original applet has a class named AgeClass.  Still to do is to rename
files, but this is an interim commit changing just the file contents,
for the purposes of traceability.