Add right-click popup to the charts
This is still the wrong chart type on the race graph.

There is also a known issue if you manage to drag out of the viewport,
release the mouse button and then move the mouse quickly back over the
chart.  In this case the popup is still drawn even though the mouse
button is no longer held down, because the mouse-up event happened out
of the viewport.  (The fix might be as simple as checking whether the
buttons are held down in the mouse-move event.)