Handle some age classes with no winners
SplitsBrowser can now handle classes with no winner, provided each
control on the course is punched by at least one competitor.  It can't
yet handle the case that a control is not punched by any competitor on
the course.

Ultimately it seems this latter reason is the cause of most winner-less
classes: a class containing only a single competitor, who mispunches.