Minor fixes to SI-reader test.
Two tests that were supposed to test the handling of competitors with no
splits were broken because the competitors that they created had finish
splits.  The finish time needed to be wiped out too.  Unfortunately, the
code was also broken, so the tests incorrectly passed.

Also, a grammatical issue in an assertion message was fixed.