Identify and replace hacky constants
The margin applied to the body element accounted for most of these
numbers.  Hiding and re-showing the chart helps to remove scrollbars, as
does subtracting off an extra pixel from the width when using IE.

Additionally, also for IE's benefit, the chart has had the style
overflow:hidden applied to it.