Improve popup behaviour in IE9/IE10
If the mouse subsequently re-enters the chart, and a mouse-out timeout
is currently in progress, cancel said timeout.  This happens more often
than might be expected, even though the timeout period is 1 millisecond.

This change fixes some odd behaviour with the popup dialog in IE9/IE10
emulation modes in IE11.