Remove control line when redrawing chart
The chart is redrawn following a change of browser window size.  This
includes zooming (e.g. using Ctrl+scrollwheel).   The control line won't
necessarily be in the same place after the zooming, we can't get the
current mouse location during the resize, and besides, as the window has
been zoomed in or out, the old location of it wouldn't be of any use to
us anyway.   So we remove the line.