Calculate time-loss statistics
The 'Time loss' statistic, as calculated by the original SplitsBrowser,
is now added.  The calculation is the same apart from a slight
simplification.

Additionally, three existing tests, which included a competitor with a
different number of split times to the number of controls on their
course, were found and the problems fixed.