Adjust layout within the top bar
The aim here was to ensure that if the top-bar controls wrap onto two
lines, then SplitsBrowser takes this into account.

This was achieved by floating everything inside it and adding a an empty
div with clear:both after everything else in it.  The clear:both styling
could then go from the main panel.