Handle certain encoding errors in IE better
IE ca make such a poor job of recovering from a single non-ASCII byte in
data that is supposedly UTF-8, that it may end up swallowing the
non-ASCII byte and a few characters that follow it.  This may include
the next semicolon in an SI CSV file.  Clearly, by the time
SplitsBrowser gets the data, it's too late to do anything about it.
Instead, we try provide a more helpful error message.