Allow for 'dubious' splits
A competitor now has 'original' as well as 'final' cumulative and split
times.  The former is read from the file and shown in the results table,
whereas the latter is what is plotted.  The intention is that some of
the splits can be marked as 'dubious' and omitted from the chart, with
dotted lines connecting the points on either side.  A split is marked as
dubious by replacing it with NaN.

This commit does not include the ability to detect any dubious splits,
nor to adjust the chart to draw dotted lines around dubious splits.