Remove warning when nobody punches a control
The code in the previous commit for handling controls that everybody on
a course has a dubious split for now handles this situation.  So there's
no need to check for this situation.

The ability to disable the chart and show a warning over it, and disable
the competitor listbox, has also gone.