Add control for showing original/repaired data
Control appears if any data was repaired.

It will be better to only make it appear if the selected class/classes
contain repaired data.  Event 6502 has broken data only in the last of
its classes (SB), and someone seeing the checkbox might wonder what is
wrong with all of the other classes.