Relax time count on old-format SI HTML files
There is no longer a check that the number of cumulative and split times
are the same in the old (preformatted) SI HTML reader, as splits can be
omitted without trace and the parser could not know.  This can happen in
particular if there are negative splits.

This issue is known to hit three events (6558, 6844, 6865).