Fix bug with fastest-time-plus-N% charts
The underlying issue was an assumption that the reference cumulative
times are integers.  For fastest-time-plus-N% charts this isn't always
true if N isn't a multiple of 100.

Replacing parseInt with parseFloat would appear to fix the problem, but
as d3.map converts the float keys to strings, I didn't want to assume
that I would always get the same number back after doing a float ->
string -> float conversion.  So instead of relying on a d3.map to
deduplicate the times I did so myself.

Fixes #17.