Read start times correctly in SI reader
The time in the 'Start punch' column will be used if it is available.
If not (for example, because the event uses a timed start rather than a
punching start), the Start Time column will be used instead.

This was already supported with the 60-column variation.  Support for it
has been tidied up and added to the other two variations.