Fix bug with chart type selector
It was possible to view the race graph for classes that you should not
be able to if you had two classes, one that could show the race graph
and one that could not, view the race graph for the class that could and
then change to the class that could not.  This has been fixed.

Additionally the chart type selector test was using a non-existent
chart-type property.  This has now been fixed.