Introduce 'data' build
The 'data' build contains only the results data reading and supported
files; it does not contain the GUI controls.  This could be useful for a
league-table calculator.

Furthermore, the data build does not use messages, so it does not need
one of SplitsBrowser's language files In order to achieve this, a method
was removed from competitor.js and moved to chart.js.