Some HTML tidying-up
The main intention is to make it easier to use SplitsBrowser with
right-to-left languages, so CSS classes are used to apply float:left or
float:right to the controls on the top layer.

Some possibly-invalid HTML has been fixed (particularly regarding child
elements of spans) - this caused some issues with elements not always
being rendered in the correct order even when RTL was in force.