Reject HTML in the CSV reader
Event 7539 on splitsbrowser.org.uk is in an unrecognised HTML format
which happens to have a single comma on the first line.  This caused the
CSV reader to think that it was CSV, which it would then fail to parse.
Now, the CSV reader checks for the presence of HTML and refuses to
handle the data if it finds any.