Fix issue with event 7632
Event 7632 is a CSV file that ends with a line 'NOCLAS,-1'.  There are
no competitors in this class, so it would seem a bit harsh to reject the
file when this section can easily be ignored.  We'll still check for
non-negative control counts, but only if there are any competitors in
the class.