Fix a couple of issues with IOF v2.0.3 XML files
There are two problems here, both caused by XML files generated by
Winsplits Pro.

Firstly, if a competitor misses a control, the Time element may be
omitted, and SplitsBrowser doesn't handle this well, especially as the
Time element isn't mandatory in the DTD.

Secondly, the competitor's total time may be inside a <Clock> element
within the <Time> element we are expecting.  jQuery(...).text() still
returns the text of the inner element, but we also end up with
whitespace before and after the time.  This goes against the DTD, but
fortunately has an easy fix: trim all times in parseTime().

Fixes #33