Create 'artificial' course IDs for IOF XML files
These artificial course IDs are formed by joining the course's list of
control codes together, which ought to identify the course: two classes
with the same controls in the same order are very likely to be using the
same course.

Additionally, IOF XML v3 files should null out the course ID if it is
not found.

Fixes #34.